+++ assets/scss/components/elements/blog/_single.scss
@@ -4,7 +4,7 @@
 .page {
 
 	.neve-main {
-		margin: var(--content-vspacing);
+		margin: var(--c-vspace);
 	}
 }
